Fix command 'make' , add local repos support, edit version
Support for local repo was added. User can add local directory as a
repo, which will be scanned. After that the manifest containing all the
packages in local_dir is created. Scanning local repos is optional.

Command 'make' was edited. Now user just types 'make' in the working
directory, which contains the package data including rockspec. Deps are
searched for in all the repos including local repos. If dependency isn't
present in the local_repo, it's fetched from the remote repo